python_backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
ultralytics
opencv-python
Pillow
numpy
pydantic
python-multipart
facenet-pytorch
dlib
Pillow